Move extension examples from package .../impl to .../example
