+++ core/src/net/sf/openrocket/simulation/extension/example/CSVSave.java
@@ -1,4 +1,4 @@
-package net.sf.openrocket.simulation.extension.impl;
+package net.sf.openrocket.simulation.extension.example;
 
 import java.io.File;
 import java.io.FileNotFoundException;
